+++ plugins/transforms/janino/src/main/java/org/apache/hop/pipeline/transforms/javafilter/JavaFilter.java
@@ -27,7 +27,7 @@ import org.apache.hop.pipeline.PipelineMeta;
 import org.apache.hop.pipeline.transform.BaseTransform;
 import org.apache.hop.pipeline.transform.ITransform;
 import org.apache.hop.pipeline.transform.TransformMeta;
-import org.apache.hop.pipeline.transform.errorhandling.IStream;
+import org.apache.hop.pipeline.transform.stream.IStream;
 import org.codehaus.janino.ExpressionEvaluator;
 
 import java.math.BigDecimal;
